src: size_t conversion test_7_8_*.

+++ src/compliance_tests/TestIso_7_8_1_1.cpp
@@ -125,14 +125,14 @@ class TestIso_7_8_1_1 : public test::TestBase
             // see the dominant flipped bits!
             brs->val_ = BitVal::Recessive;
 
-            int dominant_pulse_length;
+            size_t dominant_pulse_length;
 
             if (elem_test.index_ == 1)
                 dominant_pulse_length = nbt.prop_ + nbt.ph1_;
             else
                 dominant_pulse_length = nbt.prop_ + nbt.ph1_ + 1;
 
-            for (int j = 0; j < dominant_pulse_length; j++)
+            for (size_t j = 0; j < dominant_pulse_length; j++)
                 brs->ForceTQ(j, BitVal::Dominant);
 
             drv_bit_frm->Print(true);
